CM-1810: Moved test cases related to the experiment assets logging into LogAssetsSupportTest for better clarity.
